I have started a folder in the main source tree:
  src/graphlab/yrl

to store extensions added as part of my internship at Yahoo Labs.
This source will only be built if the "undocumented" configure option:

  ./configure --yrlabs

is provided.  I am avoiding branching to ensure that extensions stay
compatible with changes in the main trunk and to allow the rest of the
GraphLab team to easily test new features that will eventually be
moved into the main branch (pending approval). Please notify me
immediately if something infects the main build.