Fixed incompatibility issues between Boost Filesystem v2 and v3.
Configured CMake to define BOOST_FILESYSTEM_VERSION=3 if boost version is
>=1.46, and to define BOOST_FILESYSTEM_VERSION=2 otherwise. Updated fs_util.cpp
and io.cpp to use this macro. This is in accordance with the recommendations
of the Boost developers: http://www.boost.org/doc/libs/1_48_0/libs/filesystem/v3/doc/index.htm

In addition, I also updated the README to indicate GraphLab's dependency on
the Boost iostreams library.