Make sure __builtin_ia32_crc32di is properly detected
Without `clang -msse4.2 ...`, clang will not enable
the `__builtin_ia32_crc32di` intrinsic. However, because
of the way the config check for that intrinsic was written,
clang noticed that the call in the test had no observable
effects, so it optimized the call away. By returning
the result of the call, we guarantee this call will
actually be called, and it will properly error out the
test by complaining that the `__builtin_ia32_crc32di` is
an unknown symbol.