Add support for manually disabling tcmalloc
For unknown reasons tcmalloc seems to not work properly on some
mac setups. The workaround until today is to manually edit [1]
CMakeLists.txt so that GraphLab is built without linking against
tcmalloc.

This patch introduces the `--no_tcmalloc` option to the configure
script (along with the CMake option -DNO_TCMALLOC) so that such
manual edits aren't really necessary when the user decides to
build GraphLab without tcmalloc.

Fixes #55. While I agree that the best approach would be to
actually fix the underlying problem, this patch enables mac
users facing the problem to work around the current issue more
easily.

NOTE: This is better viewed with the `-w` option to ignore the
indentation change incurred by adding the if()-endif() block.

[1]: https://groups.google.com/d/msg/graphlab-kdd/SGfjqej1TqM/XTRMWg7946cJ